requirements.txt
transformers
python-dotenv
pymongo
textblob
nltk